Move `EdgeEndpointSet` to `subgraph` module
